+++ pkg/tui/list/workload/select.go
@@ -18,15 +18,9 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/charmbracelet/huh/spinner"
-	"github.com/sirupsen/logrus"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-
 	"github.com/silogen/kaiwo/pkg/k8s"
 	tuicomponents "github.com/silogen/kaiwo/pkg/tui/components"
 	podlist "github.com/silogen/kaiwo/pkg/tui/list/pod"
-	"github.com/silogen/kaiwo/pkg/workloads"
-	"github.com/silogen/kaiwo/pkg/workloads/factory"
 )
 
 func RunSelectWorkloadType(_ context.Context, _ k8s.KubernetesClients, state *tuicomponents.RunState) (tuicomponents.StepResult, tuicomponents.RunStep[tuicomponents.RunState], error) {
@@ -50,64 +44,65 @@ func RunSelectWorkloadType(_ context.Context, _ k8s.KubernetesClients, state *tu
 }
 
 func runSelectWorkload(ctx context.Context, clients k8s.KubernetesClients, state *tuicomponents.RunState) (tuicomponents.StepResult, tuicomponents.RunStep[tuicomponents.RunState], error) {
-	labelSelector := client.MatchingLabels{}
-
-	if state.User != "" {
-		labelSelector[workloads.KaiwoUsernameLabel] = state.User
-	}
-
-	var err error
-
-	var workloadReferences []workloads.Workload
-
-	loadReferences := func() {
-		workloadReferences, err = factory.ListObjects(ctx, clients.Client, state.WorkloadType, labelSelector, client.InNamespace(state.Namespace))
-	}
-
-	if spinnerErr := spinner.New().Title("Listing workloads").Action(loadReferences).Run(); spinnerErr != nil {
-		return tuicomponents.StepResultErr, nil, spinnerErr
-	}
-
-	if err != nil {
-		return tuicomponents.StepResultErr, nil, fmt.Errorf("failed to list workloads: %w", err)
-	}
-
-	columns := []string{
-		"Name",
-		"Status",
-		"Kaiwo user",
-	}
-
-	var data [][]string
-
-	dataMap := map[string]workloads.Workload{}
-
-	for _, workloadReference := range workloadReferences {
-		data = append(data, []string{
-			workloadReference.GetName(),
-			workloadReference.GetStatus(),
-			workloadReference.GetKaiwoUser(),
-		})
-		dataMap[workloadReference.GetName()] = workloadReference
-	}
-
-	if len(data) == 0 {
-		logrus.Warnf("No %s workloads found", state.WorkloadType)
-		return tuicomponents.StepResultPrevious, nil, nil
-	}
-
-	title := fmt.Sprintf("Select the %s workload", state.WorkloadType)
-	selectedRow, result, err := tuicomponents.RunSelectTable(data, columns, title, true)
-	if err != nil {
-		return result, nil, fmt.Errorf("failed to select the workload: %w", err)
-	}
-	if result == tuicomponents.StepResultOk {
-		state.Workload = workloadReferences[selectedRow]
-	} else {
-		state.Workload = nil
-	}
-
-	return result, runSelectWorkloadAction, nil
+	panic("")
+	//labelSelector := client.MatchingLabels{}
+	//
+	//if state.User != "" {
+	//	labelSelector[workloads.KaiwoUsernameLabel] = state.User
+	//}
+	//
+	//var err error
+	//
+	//var workloadReferences []workloads.Workload
+	//
+	//loadReferences := func() {
+	//	workloadReferences, err = factory.ListObjects(ctx, clients.Client, state.WorkloadType, labelSelector, client.InNamespace(state.Namespace))
+	//}
+	//
+	//if spinnerErr := spinner.New().Title("Listing workloads").Action(loadReferences).Run(); spinnerErr != nil {
+	//	return tuicomponents.StepResultErr, nil, spinnerErr
+	//}
+	//
+	//if err != nil {
+	//	return tuicomponents.StepResultErr, nil, fmt.Errorf("failed to list workloads: %w", err)
+	//}
+	//
+	//columns := []string{
+	//	"Name",
+	//	"Status",
+	//	"Kaiwo user",
+	//}
+	//
+	//var data [][]string
+	//
+	//dataMap := map[string]workloads.Workload{}
+	//
+	//for _, workloadReference := range workloadReferences {
+	//	data = append(data, []string{
+	//		workloadReference.GetName(),
+	//		workloadReference.GetStatus(),
+	//		workloadReference.GetKaiwoUser(),
+	//	})
+	//	dataMap[workloadReference.GetName()] = workloadReference
+	//}
+	//
+	//if len(data) == 0 {
+	//	logrus.Warnf("No %s workloads found", state.WorkloadType)
+	//	return tuicomponents.StepResultPrevious, nil, nil
+	//}
+	//
+	//title := fmt.Sprintf("Select the %s workload", state.WorkloadType)
+	//selectedRow, result, err := tuicomponents.RunSelectTable(data, columns, title, true)
+	//if err != nil {
+	//	return result, nil, fmt.Errorf("failed to select the workload: %w", err)
+	//}
+	//if result == tuicomponents.StepResultOk {
+	//	state.Workload = workloadReferences[selectedRow]
+	//} else {
+	//	state.Workload = nil
+	//}
+	//
+	//return result, runSelectWorkloadAction, nil
 }
 
 func runSelectWorkloadAction(_ context.Context, _ k8s.KubernetesClients, _ *tuicomponents.RunState) (tuicomponents.StepResult, tuicomponents.RunStep[tuicomponents.RunState], error) {
