--- pkg/workloads/jobs/job.yaml.tmpl
@@ -1,109 +1,0 @@
-apiVersion: batch/v1
-kind: Job
-metadata:
-  name: {{ .Meta.Name }}
-  namespace: {{ .Meta.Namespace }}
-  labels:
-    kueue.x-k8s.io/queue-name: {{ .WorkloadMeta.Queue }}
-    kaiwo-cli/username: {{ .Meta.User }}
-spec:
-  ttlSecondsAfterFinished: 3600
-  template:
-    spec:
-      securityContext:
-        runAsUser: 1000
-        runAsGroup: 1000
-        fsGroup: 1000
-      restartPolicy: "Never"
-      {{- if .Meta.ImagePullSecret }}
-      imagePullSecrets:
-      - name: {{ .Meta.ImagePullSecret }}
-      {{- end }}
-      containers:
-      - name: {{ .Meta.Name }}
-        image: {{ .Meta.Image }}
-        imagePullPolicy: Always
-        {{- if .Workload.Entrypoint }}
-        command: 
-        - sh
-        - -c
-        - {{ .Workload.Entrypoint }}
-        {{- end }}
-        env:
-          - name: HF_HOME
-            value: /workload/.cache/huggingface
-        {{- if .Meta.EnvVars }}
-        {{- range .Meta.EnvVars }}
-        {{- if .Value }}
-          - name: {{ .Name }}
-            value: "{{ .Value }}"
-        {{- else }}
-          - name: {{ .Name }}
-            valueFrom:
-              secretKeyRef:
-                name: {{ .ValueFrom.SecretKeyRef.Name }}
-                key: {{ .ValueFrom.SecretKeyRef.Key }}
-        {{- end }}
-        {{- end }}
-        {{- end }}
-        resources:
-          requests:
-            memory: "{{ mul .Scheduling.TotalRequestedGPUs 32 }}Gi"
-            cpu: "{{ mul .Scheduling.TotalRequestedGPUs 4 }}"
-            amd.com/gpu: "{{ .Scheduling.TotalRequestedGPUs }}"
-          limits:
-            memory: "{{ mul .Scheduling.TotalRequestedGPUs 32 }}Gi"
-            cpu: "{{ mul .Scheduling.TotalRequestedGPUs 4 }}"
-            amd.com/gpu: "{{ .Scheduling.TotalRequestedGPUs }}"
-        volumeMounts:
-          - mountPath: /workload
-            name: {{ .Meta.Name }}-main
-        {{- if .Meta.SecretVolumes }}
-        {{- range .Meta.SecretVolumes }}
-          - name: {{ .Name }}
-            mountPath: {{ .MountPath }}
-            subPath: {{ .SubPath }}
-        {{- end }}
-        {{- end }}
-        {{- if .Meta.HasConfigMap }}
-          - mountPath: /workload/mounted
-            name: workload-mount
-        {{- end }}
-          - mountPath: /dev/shm
-            name: dshm
-      volumes:
-      {{- if .Scheduling.Storage }}
-        - name: {{ .Meta.Name }}-main
-          ephemeral:
-            volumeClaimTemplate:
-              spec:
-                accessModes: [ "ReadWriteMany" ]
-                storageClassName: {{ .Scheduling.Storage.StorageClassName }}
-                resources:
-                  requests:
-                    storage: {{ .Scheduling.Storage.Quantity }}
-      {{- else }}
-        - name: {{ .Meta.Name }}-main
-          emptyDir:
-            medium: Memory
-            sizeLimit: 10Mi
-      {{- end }}
-      {{- if .Meta.SecretVolumes }}
-      {{- range .Meta.SecretVolumes }}
-        - name: {{ .Name }}
-          secret:
-            secretName: {{ .SecretName }}
-            items:
-              - key: {{ .Key }}
-                path: {{ .SubPath }}
-      {{- end }}
-      {{- end }}
-      {{- if .Meta.HasConfigMap }}
-        - name: workload-mount
-          configMap:
-            name: {{ .Meta.Name }}
-      {{- end }}
-        - name: dshm
-          emptyDir:
-            medium: Memory
-            sizeLimit: 200Gi
