--- pkg/workloads/utils/select.go
@@ -1,32 +1,0 @@
-// Copyright 2025 Advanced Micro Devices, Inc.  All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//       http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package utils
-
-import (
-	corev1 "k8s.io/api/core/v1"
-)
-
-type PodSelectionPredicate func(pod corev1.Pod) bool
-
-func IsGPUPod(pod corev1.Pod) bool {
-	for _, container := range pod.Spec.Containers {
-		for resourceName := range container.Resources.Limits {
-			if resourceName == "nvidia.com/gpu" || resourceName == "amd.com/gpu" {
-				return true
-			}
-		}
-	}
-	return false
-}
